pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gpx-flyover"
version = "0.2.0"
description = "Generate 3D flyover videos from GPX files"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["gpx", "video", "3d", "terrain", "cycling", "hiking", "flyover"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: GIS",
]
dependencies = [
    "gpxpy>=1.6.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "pyvista>=0.44.0",
    "matplotlib>=3.7.0",
    "SRTM.py>=0.3.7",
    "Pillow>=10.3.0",
    "requests>=2.31.0",
    "click>=8.0.0",
    "tqdm>=4.65.0",
]

[project.scripts]
gpx-flyover = "gpx_flyover.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
